Database.RethinkDB.TH

Description

Functions to mechanically derive ToDatum and FromDatum instances. Note that you need to enable the TemplateHaskell language extension in order to use this module.

An example shows how instances are generated for arbitrary data types. First we define a data type:

data D a = Nullary
         | Unary Int
         | Product String Char a
         | Record { testOne   :: Double
                  , testTwo   :: Bool
                  , testThree :: D a
                  } deriving Eq

Next we derive the necessary instances. Note that we make use of the feature to change record field names. In this case we drop the first 4 characters of every field name. We also modify constructor names by lower-casing them:

$(deriveDatum defaultOptions{fieldLabelModifier = drop 4, constructorTagModifier = map toLower} ''D)

Now we can use the newly created instances.

d :: D Int
d = Record { testOne = 3.14159
           , testTwo = True
           , testThree = Product "test" 'A' 123
           }

>>> fromDatum (toDatum d) == Success d
> True

Please note that you can derive instances for tuples using the following syntax:

-- FromDatum and ToDatum instances for 4-tuples.
$(deriveDatum defaultOptions ''(,,,))

Encoding configuration

data Options :: *

Constructors

Options

Fields

fieldLabelModifier :: String -> String

constructorTagModifier :: String -> String

allNullaryToStringTag :: Bool

omitNothingFields :: Bool

sumEncoding :: SumEncoding

data SumEncoding :: *

Constructors

TaggedObject

Fields

tagFieldName :: String

contentsFieldName :: String

ObjectWithSingleField

TwoElemArray

defaultOptions :: Options

defaultTaggedObject :: SumEncoding

FromDatum and ToDatum derivation

deriveDatum

Arguments

:: Options
Encoding options.

-> Name
Name of the type for which to generate ToDatum and FromDatum instances.

-> Q [Dec]

Generates both ToDatum and FromDatum instance declarations for the given data type.

This is a convienience function which is equivalent to calling both deriveToDatum and deriveFromDatum.

deriveToDatum

Arguments

:: Options
Encoding options.

-> Name
Name of the type for which to generate a ToDatum instance declaration.

-> Q [Dec]

Generates a ToDatum instance declaration for the given data type.

deriveFromDatum

Arguments

:: Options
Encoding options.

-> Name
Name of the type for which to generate a FromDatum instance declaration.

-> Q [Dec]

Generates a FromDatum instance declaration for the given data type.

mkToDatum

Arguments

:: Options
Encoding options.

-> Name
Name of the type to encode.

-> Q Exp

Generates a lambda expression which encodes the given data type as Datum.

mkParseDatum

Arguments

:: Options
Encoding options.

-> Name
Name of the encoded type.

-> Q Exp

Generates a lambda expression which parses the Datum encoding of the given data type.